+++ src/oldnews/data/config.py
@@ -31,9 +31,6 @@ class Configuration:
     show_all: bool = False
     """Should we show all articles, or just unread articles?"""
 
-    mark_read_on_read_timeout: float = 0.25
-    """How long to wait before marking an article as read, upon reading it."""
-
     startup_refresh_holdoff_period: float = 600
     """The number of seconds to wait before hitting TheOldReader again on startup."""